+++ targets/Win32/plc_Win32_main.c
@@ -8,7 +8,7 @@
 #include <windows.h>
 
 /* provided by POUS.C */
-extern unsigned long common_ticktime__;
+extern unsigned long long common_ticktime__;
 
 long AtomicCompareExchange(long* atomicvar, long compared, long exchange)
 {
@@ -31,7 +31,7 @@ void PLC_timer_notify()
 }
 
 HANDLE PLC_timer = NULL;
-void PLC_SetTimer(long long next, long long period)
+void PLC_SetTimer(unsigned long long next, unsigned long long period)
 {
 	LARGE_INTEGER liDueTime;
 	/* arg 2 of SetWaitableTimer take 100 ns interval*/
